+++ app/(root)/posts/[id]/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import PostDetail from "@/module/post/components/PostDetail";
+import PostDetail from "@/modules/post/components/PostDetail";
 
 const Page = async ({ params }: { params: Promise<{ id: number }> }) => {
   const id = (await params).id;
