--- module/comment/interface.ts
@@ -1,7 +1,0 @@
-export interface Comment {
-  id: number;
-  post_id: number;
-  name: string;
-  email: string;
-  body: string;
-}
