--- module/post/interface.ts
@@ -1,7 +1,0 @@
-export interface Post {
-  id?: number;
-  user_id?: number;
-  title: string;
-  body: string;
-  imageUrl?: string;
-}
