--- module/user/interface.ts
@@ -1,7 +1,0 @@
-export interface User {
-  id: number;
-  name: string;
-  email: string;
-  gender: "male" | "female";
-  status: "active" | "inactive";
-}
